+++ CHANGELOG.md
@@ -18,6 +18,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 * Cleaned up predicted vs. actual code for predictors
 * HelmRuns.coerce can now accept a more expressive set of inputs
 * BREAKING: You must how specify `helm_runs` when calling the predictor.
+* `magnet download helm` can now download multiple benchmarks
 
 ### Fixed